+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Unreleased
 
+- Major: Add combat achievement notifications. (#63)
 - Major: Add additional information for PvP deaths. (#55, #58)
 - Major: Add boss kill count notifications. (#50)
 - Minor: Add config to ignore slayer tasks with low point values. (#61)
+- Dev: Add plugin-hub icon. (#65)
 
 ## 1.0.3